Remove subcommand `kubectl mongodb debug`
There are two ways to collect diagnostic information from customers enviornments. Either
we can ask them to run `kubectl mongodb debug` or we can ask them to run `public/support/mdb_operator_diagnostic_data.sh`
script.
From the field we got to know that people prefer .sh over the binary, that's why it looks
like we have been providing the .sh file to the users.
And because of that having `kubectl mongodb debug` just confuses us as to, which one should be
provided to the users. That's why this commit removes that subcommand.