[CLOUDP-333407] Remove subcommand `kubectl mongodb debug` (#293)
# Summary

There are two ways to collect diagnostic information from customers
enviornments. Either we can ask them to run `kubectl mongodb debug` or
we can ask them to run `public/support/mdb_operator_diagnostic_data.sh`
script.
From the field we got to know that people prefer .sh over the binary,
that's why it looks like we have been providing the .sh file to the
users. And because of that having `kubectl mongodb debug` just confuses
us as to, which one should be provided to the users. That's why this
commit removes that subcommand.

## Proof of Work

Build the binary and make sure that `debug` subcommand has been removed.

```
./kubectl-mongodb                                                       
This application is a tool to simplify maintenance tasks
of MongoDB resources in your kubernetes cluster.

Build: 709d41f97a48f8e1970707401e4c8172ce163eeb, 2025-07-24T08:00:11Z

Usage:
  kubectl-mongodb [command]

Available Commands:
  completion   Generate the autocompletion script for the specified shell
  help         Help about any command
  multicluster Manage MongoDB multicluster environments on k8s

Flags:
  -h, --help   help for kubectl-mongodb

Use "kubectl-mongodb [command] --help" for more information about a command.
```

I tried to verify if we ever documented this, to make sure if we will
have to change the docs or not. And it looks like this subcommand was
not documented
https://www.mongodb.com/docs/kubernetes/current/plugin-reference/. So we
will not have to create/push release notes for this.

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you checked for release_note changes?