CLOUDP-331886 - add new agent binaries as base (#292)
# Summary

- those scripts are required for agent matrix removal. They are placed
here - since we overwrite the tag and this would race with the images
created in other e2e tests

### Dockerfile updates:
* Updated `docker/mongodb-agent-non-matrix/Dockerfile` to copy new
scripts (`agent-launcher-shim.sh`, `setup-agent-files.sh`,
`dummy-probe.sh`, `dummy-readinessprobe.sh`) into the container at build
time. These scripts are placed in `/usr/local/bin` for execution.
* Updated `docker/mongodb-agent-non-matrix/Dockerfile.builder` to
include the same script files in the build process, ensuring they are
available in the image.

### New scripts added:
*
[`agent-launcher-shim.sh`](diffhunk://#diff-7050c43142cef16fa7268fd7d1356e96bdf5de2c8bc0d050f277babee01a397cR1-R31):
A script to initialize the agent launcher by setting up agent files and
starting the launcher if the required script is found.
*
[`setup-agent-files.sh`](diffhunk://#diff-fbe2cd71b8fc58242e04e18562c58b1085f6b059dce42d94a2f4e5cbdf7a3469R1-R95):
A script to set up dummy probe scripts initially, wait for an init
container to provide real probe and launcher scripts, and replace the
dummy scripts with the real ones.
*
[`dummy-probe.sh`](diffhunk://#diff-ff81b0610fd08dd2d32bcbe72c0d78fb402d5f6dbe8a388bad94c7cdf9b95e3bR1-R5):
A dummy liveness probe script that always returns success to keep the
container alive during the setup process.
*
[`dummy-readinessprobe.sh`](diffhunk://#diff-4378f36df04937eb7473771a47dc6ac5a424ec8ffdb171e201b9fc8052a54c26R1-R5):
A dummy readiness probe script that always returns a "not ready" status
until the real probe script is available.

## Proof of Work

- not used anywhere, ci needs to pass
- used by this pr (that contains more information on how its used):
https://github.com/mongodb/mongodb-kubernetes/pull/267

## Checklist
- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?