Fix shellcheck issues in the kubectl-mongodb plugin scripts (#297)
# Summary

This RP fixes the shellcheck (info) issues that we saw in the
`kubectl-mongodb` plugin related scripts when we ran `make precommit`.
Most of the issues are fixed when I didn't actually have to make any
change but I had to disable check for some lines where it was expected
to make some change and I was not sure if that would impact things
functionally.

## Proof of Work

NA