Correct `kubectl-mongodb` path for command `prepare_local_e2e` (#315)
# Summary

When we change the directory of `kubectl-mongodb` from
`public/tools/multicluster` to `cmd/kubectl-mongodb` we change the
source file `build_multi_cluster_kubeconfig_creator.sh` to use the new
path but I missed updating the new path in the file
`funcs/multicluster`. This PR fixes that problem and updates the path to
point to the correct location.

## Proof of Work

Run `make prepare_local_e2e`, that will build the `kubectl-mongodb`
binary and use the `kubectl mongodb multicluster` command to setup the
multicluster environment.