Fix master builds (#331)
# Summary

Master patches are broken because they build with `latest` tag only,
while the patch still need the `patch_id` tagged images.
This PR fixes this issue to get master green again, but remove builds to
`latest`.
They will be added again in
https://github.com/mongodb/mongodb-kubernetes/pull/317, using
build_info.
Not pushing to latest only impacts local testing for dev, as we use them
to have the most recently built versions from master. It won't break
local development, but it means the image used will be slightly
outdated.

## Proof of Work
I tested it by forcing to return the patch ID and dumping the relevant
variables in this patch:

https://spruce.mongodb.com/version/6899d033f94ad4000694cf89/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

But the relevant environment variable (`is_patch` being false, as
described in the added documentation) will be set to the right value
only when merging to master.