Add change log for unavailability of OM container images (#332)
# Summary

Some of the versions of ops manager container images are not
built/published by us, because of an issue that we have been facing
while updating the OM version in our project.
Those issues are fixed and like suggested
[here](https://github.com/mongodb/mongodb-kubernetes/pull/324#pullrequestreview-3105253213)
this PR adds change log to inform users that some version of this image
are unavailable.