Update `controller-gen` from `0.15.0` to `0.18.0`
After updating the k8s.io related mods when we tried to generate the
manifests using make manifests it failed with

```
/Users/vivek.s/go/pkg/mod/k8s.io/api@v0.31.11/core/v1/doc.go:21:1: missing argument "" (at <input>) /Users/vivek.s/go/pkg/mod/k8s.io/api@v0.31.11/core/v1/doc.go:21:1: missing argument "" (at <input>) k8s.io/api/core/v1:-: unknown type "k8s.io/api/core/v1".LocalObjectReference
```

and other similar issues. It simply meant that the controller-gen version is not compatible with the k8s.io/api version that we  have upgraded it to.
That's why this commit updates the controller-gen version to 0.18.0.