DEVPROD-19870: Downsize instances where possible (#349)
Ticket: [DEVPROD-19870](https://jira.mongodb.org/browse/DEVPROD-19870)
This PR is for cutting instance sizes to use high memory variants. This
should result in ~ 15k savings in a year.

**Proof of Work**

Here are the successful patches for the variants which were downsized:
**e2e_om60_kind_ubi**

[patch](https://spruce.mongodb.com/version/68a39abace035a00085936d9/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC&variant=%5Ee2e_om60_kind_ubi%24)
This results in the variant being 3-4% slower (considering total
duration not wall time) on an average.
 The longest task went from 
[47m
58s](https://spruce.mongodb.com/version/mongodb_kubernetes_f0df6f42547f5602c5c4ab120d1d7e3d0db05797/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om60_kind_ubi%24)
to [48m
53s](https://spruce.mongodb.com/version/68a39abace035a00085936d9/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om60_kind_ubi%24),
so the wall clock time change would be ~2%. The cost savings would be
~30-32%.

**e2e_om70_kind_ubi**

[patch](https://spruce.mongodb.com/version/68a39abace035a00085936d9/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC&variant=%5Ee2e_om70_kind_ubi%24)
This results in the variant being 3-4% slower (considering total
duration not wall time) on an average.
 The longest task went from 
[48m
12s](https://spruce.mongodb.com/version/mongodb_kubernetes_f0df6f42547f5602c5c4ab120d1d7e3d0db05797/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om70_kind_ubi%24)
to [47m
8s](https://spruce.mongodb.com/version/68a39abace035a00085936d9/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om70_kind_ubi%24),
so the wall clock time shouldn't change by much. The cost savings would
be ~32%.


**e2e_om80_kind_ubi**

[patch](https://spruce.mongodb.com/version/68a39abace035a00085936d9/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC&variant=%5Ee2e_om80_kind_ubi%24)
This results in the variant being 3-4% slower (considering total
duration not wall time) on an average.
 The longest task went from 
[45m
2s](https://spruce.mongodb.com/version/mongodb_kubernetes_f0df6f42547f5602c5c4ab120d1d7e3d0db05797/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om80_kind_ubi%24)
to [45m
51s](https://spruce.mongodb.com/version/68a39abace035a00085936d9/task-duration?page=0&sorts=STATUS%3AASC%3BDURATION%3ADESC&variant=%5Ee2e_om80_kind_ubi%24),
so the wall clock time should change be 1-2%. The cost savings would be
~30-32%.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details