Fix tools in agent images (#376)
# Summary

This PR fixes an issue with the new agent dockerfiles where the db tools
where not placed correctly. The binaries should be present in `/tools`.
However, our tests were still passing. For that reason, the agent-script
will now fail if the binaries are missing.

## Proof of Work

The tests should fail if the binaries are missing. Example:
https://spruce.mongodb.com/version/68af05cace35b000077fae05/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC
The pipeline should pass.

## Checklist

- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details