Improve the message in `check_precommit.sh` for go mod licenses case (#381)
# Summary

If we (MCK) started consuming a new go mod deps and because of that we
would need to regenerate (`update_licenses.sh`) the
`LICENSE-THIRD-PARTY` file for the main project as well as
`kubectl-mongodb` plugin. If we forget doing that, the CI would fail
with below error

```
We have files that differ after running pre-commit, please run the pre-commit locally
```

and running `pre-commit` doesn't really update the licenses file, to do
that, we will have to run `make precommit-with-licenses`. And that's why
this PR updates above message.

## Proof of Work

NA