Update `kind` and `kindest/node` image to latest available versions (`1.33.2`) (#367)
# Summary

As part of the release process, we update the kind and `kindest/node` to
latest available versions. This PR does that.

On EVG environments the script `setup_kind_cluster.sh` tries to pull the
image from the ecr mirror instead of pulling it from docker hub to avoid
rate limiting issues. That's why to make sure that CI passes on this PR,
we will have to copy the image manually to the ecr from docker hub.
Below command can be used to do that

```
# use index digest
docker pull kindest/node:v1.33.2@sha256:c55080dc5be4f2cc242e6966fdf97bb62282e1cd818a28223cf536db8b0fddf4
skopeo copy --preserve-digests --all docker://docker.io/kindest/node:v1.33.2 docker://268558157000.dkr.ecr.eu-west-1.amazonaws.com/docker-hub-mirrors/kindest/node:v1.33.2
```
The index digest on ecr will be same and that's the same digest that can
be used in the script.

After this change somehow the test
`e2e_static_multi_cluster_kind/e2e_multi_cluster_replica_set_migration`
started failing mainly because of timeout waiting for sts to be ready.
And we figured that increasing the timeout is the best path forward.
That's why we are also increasing timeout for that test as part of this
PR.

## Proof of Work

Successful CI run on this PR would prove that the change is good.