CLOUDP-342319 - Fix proxy env vars (#398)
# Summary

This PR fixes a bug where the mongodb agents were not using the
`NO_PROXY` environment variable set on the operator. This is an issue
with the agent where setting the `httpProxy` flag will ignore the
environment variables. Therefore, running the agent without that flag
work properly whether the variables are set or not.

To make sure this is tested, the `e2e_operator_proxy` test was updated
by adding a `NO_PROXY` variable set to cloud-qa and asserting that the
proxy does not intercept those calls.

## Proof of Work

Ran a
[patch](https://spruce.mongodb.com/version/68b6ca1169f9af0007311434/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)
where the test was updated, but the fix was not added. It failed,
proving that the setting the `NO_PROXY` variable did not work.

With the fix, the CI should be green.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details