Fix `.gitignore` to not ignore `mongodb-kubernetes` dir/file
As part of this PR https://github.com/mongodb/mongodb-kubernetes/pull/21 we changed
all the occurrences of old repo to the new repo and in doing so we ended up
adding `mongodb-kubernetes` to the `.gitignore` file.

Becauase of this, if we try to add any file to our new repo in dir `public/dockerfiles/mongodb-kubernetes`
it would be ignored by default. This commit fixes that. We don't really need that in .gitignore.