Fix `.gitignore` to not ignore `mongodb-kubernetes` dir/file (#410)
# Summary

As part of this PR https://github.com/mongodb/mongodb-kubernetes/pull/21
we changed all the occurrences of old repo to the new repo and in doing
so we ended up adding `mongodb-kubernetes` to the `.gitignore` file.

Becauase of this, if we try to add any file to our new repo in dir
`public/dockerfiles/mongodb-kubernetes` it would be ignored by default.
This PR fixes that. We don't really need that in .gitignore.

## Proof of Work

Create a file in the dir `public/dockerfiles/mongodb-kubernetes` and
make sure that it shows up in `git status`.