Update the CloudManager version to `13.40.1.9747-1`
This bump PR was raised last Monday and then we figured out there
is a problem in CM version.
A new version of CM was provided by the team later to test things, but
since this PR was raised before, it has old version of CM.
This commit just updates the version to the one that was provided to us
to test things, to see if that fixes the issues.