Disable `e2e_om_ops_manager_queryable_backup` in OM8 e2e variant (#466)
This pull request introduces a temporary workaround in the
`.evergreen.yml` configuration to address an issue with the queryable
backup task in OM8. A new task group is added that excludes the
problematic task, and the build variant is updated to use this new group
until the issue is resolved in: https://jira.mongodb.org/browse/CLOUDP-346008.