[CLOUDP-343558] Run helm chart-testing tool `ct` as part of CI (`lint_repo`) (#450)
# Summary

Whenever we raised PR in to helm chart repo CI failed there because we
had linting issues in our helm chart. A previous PR that [I
raised](https://github.com/mongodb/mongodb-kubernetes/pull/445) fixed
the linting problem. But we should have a linting mechanism in our CI
that would consistently check if we have made any mistake in our charts.
This PR does that.

The `chart_schema.yaml` and `lintconf.yaml` are the default yaml files
using which `ct` should be run. And `ct` is the tool that is used for
linting in our helm charts repo.

## Proof of Work

Passing `lint_repo` task here
https://spruce.mongodb.com/version/68c43aae1d5d5d000770c624/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC
and in this PR as well.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details