Fix switch-context issues + context cleanup (#508)
# Summary

The main focus of PR is changing how the `switch_context` function is
called. It is removed from most of the functions in
`.evergreen-functions.yml` (setup_building_host,
setup_building_host_minikube) and instead added directly to task and
group variables in `.evergreen.yml`. This allows us to not create a
synthetic and empty context file for each task in evergreen. Context
file is mostly needed for e2e tasks and not for building images or
preflight tasks for example.

Other changes:

* Removed the `preflight_om_image` task from `.evergreen-tasks.yml`,
which is was unnecessary - we run all preflights on master merges and
PRs.
* Additionally, several unused context files were deleted from the
repository.

## Proof of Work

Passing CI + tested staging in different PR
[run](https://spruce.mongodb.com/version/68e65607b6bd2000074ad206/tasks?page=0&sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC&statuses=failed,setup-failed,started,will-run),
because staging builds are not yet supported.

## Checklist

- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details