Recreate kind cluster on every evg host reboot  (#32)
# Summary

Evergreen hosts reboot every day or weekend (depending on your
configuration). After every reboot, inter-cluster connectivity might be
broken for some reason. Recreating the clusters is the only solution we
have so far.
This is a systemd service that runs on every boot and recreates all
clusters (including the kind-kind for single cluster tests).
Our tunnel command will now also get the kubeconfig from the host
otherwise the tunnel won't open to the (new) ports of the recreated
clusters.

The systemd service is not created by default. This needs to be
explicitly done by running
```
evg_host.sh configure --auto-reboot
```

Remove the architecture flag, and instead left it to be inferred from
`uname`.

## Proof of Work

Tested locally, but it would be nice if someone can checkout this branch
and try it for themselves.

## Checklist
- [ ] Have you linked a jira ticket and/or is the ticket in the title?
- [ ] Have you checked whether your jira ticket required DOCSP changes?
- [ ] Have you checked for release_note changes?