[CLOUDP-350669] Publish helm chart to OCI registry for dev/staging workflows (#507)
# Summary

In an effort to start using OCI compatible container registries as our
helm chart repos, this PR makes the change package and publish our helm
chart to ECR registry. A new task `publish_helm_chart` is introduced in
the build variant `init_test_run` which packages and publishes the
chart.

The details about the versioning of the chart can be found in the
document
[here](https://docs.google.com/document/d/1eJ8iKsI0libbpcJakGjxcPfbrTn8lmcZDbQH1UqMR_g/edit?tab=t.gg5ble8qlesq).

There were two wrappers (shell scripts) introduced as part of this and
the reason was we wanted to call the python scripts
(`helm_registry_login.py` and `publish_helm_chart.py`) from
`.evergreen-function.yaml` passing an argument that's build scenario.
Build scenario is already set as env var in the evg host, but there
isn't a way to do this, in other words, we couldn't do below from the
`.evergreen-functions.yaml` file

```
binary: binary: run_python.sh scripts/release/helm_registry_login.sh $build_scenario_that_is_set_as_env_on_host
``` 

And because of that we had to write the wrapper that would call the
python script with correct argument after reading the env var.


## Proof of Work


[patch](https://spruce.mongodb.com/task/mongodb_kubernetes_init_test_run_publish_helm_chart_patch_749c1a3f3302ef2bc3f70ed683dd2e522bd94bef_68e91082554272000716941c_25_10_10_13_56_20/logs?execution=0)
and
[staging](https://spruce.mongodb.com/task/mongodb_kubernetes_init_test_run_publish_helm_chart_patch_749c1a3f3302ef2bc3f70ed683dd2e522bd94bef_68e91366a609a6000785b403_25_10_10_14_08_40/logs?execution=0)
evg runs. Notice at the end of the log that dev workflow is pushing to
diff repository and staging to different.

Successful run of `init_test_run` variant in this PR would also work. 

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details