[CLOUDP-338093] Use `goreleaser` to build `kubectl-mongodb` plugin for dev/staging workflows (#320)
# Summary

Per current implementation when we release the binary we build it via
goreleaser but in our dev and staging workflows we used
`scripts/evergreen/build_multi_cluster_kubeconfig_creator.sh` (which
eventually run `go build`) to build the binary. We needed to change this
to make sure that we are using the same mechanism to build the binary in
dev/staging workflows that we use while releasing. In order to achieve
that, this PR adds functionality to build the `kubectl-mongodb` plugin
using goreleaser in dev/staging workflows.

As part of the new workflow, `build_multi_cluster_binary` evergreen
function has been changed to run `goreleaser build` and upload the
artifacts to a S3 location. After that the linux/amd64 distribution of
the binary is downloaded locally at the path where tests image expects
it.

Having the binary in the S3 bucket would eventually help us to just
promote the binaries while the release process but we are not able to
achieve that for now, because for dev/staging workflows we are not able
to notarize the generated binaries.

After this PR, potentially we can even consider just getting rid of the
script `scripts/evergreen/build_multi_cluster_kubeconfig_creator.sh`.

## Proof of Work

Run the local patch using 

```
evergreen patch --path .evergreen.yml -p mongodb-kubernetes -v init_test_run -t build_test_image -f -y -u --browse -d "Testing build of kubectl plugin using build scenario"
```
and make sure that test image is built successfully and the goreleaser
artifacts are uploaded to the S3 bucket (`mongodb-kubernetes-dev`)
correctly.

CI passing on this PR also proves that the binary is successfully built
and copied to the test image.

## Checklist

- [x] Have you linked a jira ticket and/or is the ticket in the title?
- [x] Have you checked whether your jira ticket required DOCSP changes?
- [x] Have you added changelog file?
    - use `skip-changelog` label if not needed
- refer to [Changelog files and Release
Notes](https://github.com/mongodb/mongodb-kubernetes/blob/master/CONTRIBUTING.md#changelog-files-and-release-notes)
section in CONTRIBUTING.md for more details